backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==24.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
Django==4.1
django-cors-headers==4.3.1s\backend>
django-debug-toolbar==4.3.0
django-websocket-request==0.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
ez_setup==0.9
filelock==3.13.1
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
fonttools==4.47.0
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2023.4
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7

